Refine run loop cleanup and streaming retry removal
